requirements.txt
nvsmi==0.4.2
pyserial==3.5
